+++ rollup.config.js
@@ -2,7 +2,6 @@ import fs from 'fs';
 import path from 'path';
 import babel from 'rollup-plugin-babel';
 import json from 'rollup-plugin-json';
-import { uglify } from 'rollup-plugin-uglify';
 
 // reads package.json of packages (melody-*) in packages directory
 const pkg = fs.readFileSync(path.join(process.cwd(), './package.json'));
@@ -15,7 +14,7 @@ const config = {
     output: [
         {
             file: path.resolve(process.cwd(), pkgJSON.main),
-            format: 'cjs',
+            format: 'es',
         },
     ],
     plugins: [
@@ -31,29 +30,7 @@ const config = {
     ],
 };
 
-/**
- * ES module output configuration
- */
-const esmModuleOutput = function() {
-    return {
-        file: path.resolve(process.cwd(), pkgJSON['jsnext:main']),
-        format: 'es',
-    };
-};
-
-// Let's skip ES Modules in production environment
-if (pkgJSON['jsnext:main'] && process.env.NODE_ENV !== 'production') {
-    config.output.push(esmModuleOutput());
-}
-
-if (process.env.NODE_ENV === 'production') {
-    config.plugins.push(
-        uglify({
-            mangle: {
-                toplevel: true,
-            },
-        })
-    );
-}
+// PLACEHOLDER: block to add rules for production environment
+// if (process.env.NODE_ENV === 'production') { }
 
 export default config;