+++ testsetup/melody-transform.js
@@ -1,8 +1,11 @@
+// eslint-disable-next-line
+require = require('esm')(module /*, options*/);
+
 const { transformer } = require('../packages/melody-jest-transform');
 
 // melody plugins
 const { extension } = require('../packages/melody-extension-core');
-const idom = require('../packages/melody-plugin-idom');
+const idom = require('../packages/melody-plugin-idom').default;
 
 const customConfig = {
     plugins: [extension, idom],